backend/requirements.txt
flask>=2.0
flask-cors>=4.0
pillow>=9.0
numpy>=1.21
opencv-python>=4.5

